pyproject.toml
[project]
name = "taskfit-api"
version = "0.1.0"
description = "TaskFit API - 기업 맞춤형 실무 과제 체험 플랫폼 백엔드"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "google-genai>=1.0.0",
    "google-cloud-secret-manager>=2.20.0",
    "cloud-sql-python-connector[asyncpg]>=1.12.0",
    "authlib>=1.4.0",
    "httpx>=0.28.0",
    "pyjwt[crypto]>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
